// src/types.tsx
export type User = {
  _id?: string;
  userId?: string;
  firstName: string;
  lastName: string;
  wallet?: Wallet;
  email: string;
  phone: string;
  password: string;
  birthday: string;
  home_address?: string;
  office_address?: string;
  primary_interest?: string;
  current_challenges?: string;
  is_demo: number;
  preferred_date?: string;
  last_login: Date;
  date_joined: Date;
  suspended?: boolean;
  del_falg: number;
  verified: boolean;
  salt: string;
  role: any;
  otp: number;
  otpExpiryTime: Date;
  profile_pic?: string;

  title?: "Mr." | "Ms." | "Mrs." | "Dr." | "Prof.";
  gender?: "Male" | "Female" | "Other";
  religion?: string;
  place_of_birth?: string;
  current_parish?: string;
  national_id_number?: string;
  national_id_photo?: string;
  home_location_map?: string;
  district_of_birth?: string;
  parish_of_birth?: string;
  village_of_birth?: string;
  marital_status?: "Single" | "Married" | "Divorced" | "Widowed";
  occupation?: string;
  job_title?: string;
  next_of_kin_name?: string;
  next_of_kin_phone?: string;
  next_of_kin_email?: string;
  next_of_kin_national_id?: string;
  monthly_income_level?:
    | "Less than UGX 1,000,000"
    | "UGX 1,000,000 - 5,000,000"
    | "UGX 5,000,000 - 15,000,000"
    | "Above UGX 15,000,000";
  bank_name?: string;
  bank_account_number?: string;
  bank_mobile_account?: string;
  bank_email?: string;
  highest_education_level?:
    | "Secondary (Ordinary Level)"
    | "Secondary (Advanced Level)"
    | "Tertiary"
    | "University"
    | "Other (Specify)";
  employment_status?: "Employed" | "Self-employed" | "Unemployed" | "Retired";
  current_work_address?: string;
  employer_name?: string;
  current_salary?: string;
  side_hustle_income?: string;
  securityQuestions: {
    question: string;
    answer: string;
  }[];
  is_complete: boolean;
  socketId?: string;
  isActive?: boolean;
  createdAt?: Date;
  id: string;
  active_status?: "online" | "offline" | "do not disturb" | "idle";
};
// ...
export type GroupTypes = {
  _id: string;
  name?: string;
  group_name: string;
  group_type: string;
  group_picture: string;
  description: string;
  created_by: any;
  del_flag: number;
  createdAt: Date;
  memberCount: number;
  members: User[];
  isSacco?: boolean;
  has_bf?: boolean;
  invite_link: string;
  group_state: string;
  upgraded: boolean;
  tags: string;
  suspended?: boolean;
};
// ...
export type Reaction = {
  _id?: string;
  emoji?: string;
  user_id?: string;
};

export type Message = {
  _id?: string;
  sender?: User;
  receiver_id?: User[];
  chatroom?: string;
  content?: string;
  messageType?: "text" | "image" | "video" | "sticker" | "gif";
  attachmentUrls?: {
    type: string;
    url: string;
    name: string;
  }[];
  read?: boolean;
  status?: "sending" | "sent" | "delivered" | "seen";
  reactions?: Reaction[];
  replyingTo?: Message;
  edited?: boolean;
  editedAt?: Date;
  createdAt?: Date;
  sender_id?: string;
  pinned?: boolean;
};
// ...
export interface Wallet {
  walletAddress: string;
  balance: number;
  transactionHistory: Transaction[];
}